Fix knex not escaping it's queries properly: fixes #10
As per tgriesser/knex#828, calling `toString()` on knex's postgres queries results in improper escaping. Pending that getting fixed, and perhaps a more robust integration going forward for knex->gatepost, this change integrates knex with pg.client the same way that knex does internally; by passing the query + bindings along to postgres, rather than a fully interpolated query.

Not sure if this is the perfect code to solve it, but it does solve it.